refactor(web): deduplicate TOC components
