--- apps/web/src/app/blog/_components/horizontal-toc.tsx
@@ -1,115 +1,0 @@
-"use client";
-
-import { cn } from "@sgcarstrends/ui/lib/utils";
-import { useEffect, useState } from "react";
-
-interface TocItem {
-  id: string;
-  text: string;
-  level: number;
-}
-
-export const HorizontalTOC = () => {
-  const [headings, setHeadings] = useState<TocItem[]>([]);
-  const [activeId, setActiveId] = useState<string>("");
-
-  // Extract headings from the article on mount
-  useEffect(() => {
-    const article = document.querySelector("article");
-    if (!article) return;
-
-    const elements = article.querySelectorAll("h2");
-    const items: TocItem[] = Array.from(elements)
-      .filter((el) => el.id) // Only include elements with IDs
-      .map((el) => ({
-        id: el.id,
-        text: el.textContent || "",
-        level: 2,
-      }));
-
-    setHeadings(items);
-
-    // Set initial active heading
-    if (items.length > 0) {
-      setActiveId(items[0].id);
-    }
-  }, []);
-
-  // Track scroll position and update active heading
-  useEffect(() => {
-    if (headings.length === 0) return;
-
-    const observer = new IntersectionObserver(
-      (entries) => {
-        const visibleEntries = entries.filter((entry) => entry.isIntersecting);
-        if (visibleEntries.length > 0) {
-          const topEntry = visibleEntries.reduce((prev, current) =>
-            prev.boundingClientRect.top < current.boundingClientRect.top
-              ? prev
-              : current,
-          );
-          setActiveId(topEntry.target.id);
-        }
-      },
-      {
-        rootMargin: "-80px 0px -80% 0px",
-        threshold: 0,
-      },
-    );
-
-    for (const { id } of headings) {
-      const element = document.getElementById(id);
-      if (element) {
-        observer.observe(element);
-      }
-    }
-
-    return () => observer.disconnect();
-  }, [headings]);
-
-  if (headings.length === 0) {
-    return null;
-  }
-
-  return (
-    <nav className="mb-12 border-foreground border-b-2 pb-6">
-      <div className="mb-4 font-bold text-foreground/60 text-xs uppercase tracking-[0.3em]">
-        In This Report
-      </div>
-      <div className="flex flex-wrap gap-x-6 gap-y-2">
-        {headings.map((heading, idx) => (
-          <a
-            key={heading.id}
-            href={`#${heading.id}`}
-            onClick={(e) => {
-              e.preventDefault();
-              const element = document.getElementById(heading.id);
-              if (element) {
-                element.scrollIntoView({ behavior: "smooth" });
-                setActiveId(heading.id);
-              }
-            }}
-            className={cn(
-              "group flex items-center gap-2 font-bold text-sm underline-offset-4 transition-colors hover:underline",
-              activeId === heading.id
-                ? "text-foreground"
-                : "text-foreground hover:text-foreground",
-            )}
-          >
-            <span
-              className={cn(
-                "text-xs transition-opacity",
-                activeId === heading.id
-                  ? "text-primary"
-                  : "text-primary/60 group-hover:text-primary",
-              )}
-            >
-              {String(idx + 1).padStart(2, "0")}
-            </span>
-            <span>{heading.text}</span>
-          </a>
-        ))}
-      </div>
-    </nav>
-  );
-};
